Conflicts:
	plugins/new_user_identity/config.inc.php.dist
	plugins/new_user_identity/new_user_identity.php
	plugins/new_user_identity/package.xml

+++ program/lib/Roundcube/rcube_cache.php
@@ -1,6 +1,6 @@
 <?php
 
-/*
+/**
  +-----------------------------------------------------------------------+
  | This file is part of the Roundcube Webmail client                     |
  | Copyright (C) 2011, The Roundcube Dev Team                            |
@@ -17,7 +17,6 @@
  | Author: Aleksander Machniak <alec@alec.pl>                            |
  +-----------------------------------------------------------------------+
 */
-
 
 /**
  * Interface class for accessing Roundcube cache
@@ -42,6 +41,8 @@ class rcube_cache
     private $ttl;
     private $packed;
     private $index;
+    private $debug;
+    private $index_changed = false;
     private $cache         = array();
     private $cache_changes = array();
     private $cache_sums    = array();
@@ -65,17 +66,19 @@ class rcube_cache
         $type  = strtolower($type);
 
         if ($type == 'memcache') {
-            $this->type = 'memcache';
-            $this->db   = $rcube->get_memcache();
+            $this->type  = 'memcache';
+            $this->db    = $rcube->get_memcache();
+            $this->debug = $rcube->config->get('memcache_debug');
         }
         else if ($type == 'apc') {
-            $this->type = 'apc';
-            $this->db   = function_exists('apc_exists'); // APC 3.1.4 required
+            $this->type  = 'apc';
+            $this->db    = function_exists('apc_exists'); // APC 3.1.4 required
+            $this->debug = $rcube->config->get('apc_debug');
         }
         else {
             $this->type  = 'db';
             $this->db    = $rcube->get_dbh();
-            $this->table = $this->db->table_name('cache');
+            $this->table = $this->db->table_name('cache', true);
         }
 
         // convert ttl string to seconds
@@ -87,7 +90,6 @@ class rcube_cache
         $this->packed    = $packed;
         $this->prefix    = $prefix;
     }
-
 
     /**
      * Returns cached value.
@@ -105,7 +107,6 @@ class rcube_cache
         return $this->cache[$key];
     }
 
-
     /**
      * Sets (add/update) value in cache.
      *
@@ -115,10 +116,8 @@ class rcube_cache
     function set($key, $data)
     {
         $this->cache[$key]         = $data;
-        $this->cache_changed       = true;
         $this->cache_changes[$key] = true;
     }
-
 
     /**
      * Returns cached value without storing it in internal memory.
@@ -136,7 +135,6 @@ class rcube_cache
         return $this->read_record($key, true);
     }
 
-
     /**
      * Sets (add/update) value in cache and immediately saves
      * it in the backend, no internal memory will be used.
@@ -151,7 +149,6 @@ class rcube_cache
         return $this->write_record($key, $this->serialize($data));
     }
 
-
     /**
      * Clears the cache.
      *
@@ -164,7 +161,6 @@ class rcube_cache
         // Remove all keys
         if ($key === null) {
             $this->cache         = array();
-            $this->cache_changed = false;
             $this->cache_changes = array();
             $this->cache_sums    = array();
         }
@@ -189,7 +185,6 @@ class rcube_cache
         $this->remove_record($key, $prefix_mode);
     }
 
-
     /**
      * Remove cache records older than ttl
      */
@@ -197,15 +192,14 @@ class rcube_cache
     {
         if ($this->type == 'db' && $this->db && $this->ttl) {
             $this->db->query(
-                "DELETE FROM ".$this->table.
-                " WHERE user_id = ?".
-                " AND cache_key LIKE ?".
-                " AND expires < " . $this->db->now(),
+                "DELETE FROM {$this->table}".
+                " WHERE `user_id` = ?".
+                " AND `cache_key` LIKE ?".
+                " AND `expires` < " . $this->db->now(),
                 $this->userid,
                 $this->prefix.'.%');
         }
     }
-
 
     /**
      * Remove expired records of all caches
@@ -215,19 +209,14 @@ class rcube_cache
         $rcube = rcube::get_instance();
         $db    = $rcube->get_dbh();
 
-        $db->query("DELETE FROM " . $db->table_name('cache') . " WHERE expires < " . $db->now());
+        $db->query("DELETE FROM " . $db->table_name('cache', true) . " WHERE `expires` < " . $db->now());
     }
-
 
     /**
      * Writes the cache back to the DB.
      */
     function close()
     {
-        if (!$this->cache_changed) {
-            return;
-        }
-
         foreach ($this->cache as $key => $data) {
             // The key has been used
             if ($this->cache_changes[$key]) {
@@ -241,9 +230,10 @@ class rcube_cache
             }
         }
 
-        $this->write_index();
+        if ($this->index_changed) {
+            $this->write_index();
+        }
     }
-
 
     /**
      * Reads cache entry.
@@ -260,11 +250,27 @@ class rcube_cache
         }
 
         if ($this->type != 'db') {
-            if ($this->type == 'memcache') {
-                $data = $this->db->get($this->ckey($key));
+            $this->load_index();
+
+            // Consistency check (#1490390)
+            if (!in_array($key, $this->index)) {
+                // we always check if the key exist in the index
+                // to have data in consistent state. Keeping the index consistent
+                // is needed for keys delete operation when we delete all keys or by prefix.
             }
-            else if ($this->type == 'apc') {
-                $data = apc_fetch($this->ckey($key));
+            else {
+                $ckey = $this->ckey($key);
+
+                if ($this->type == 'memcache') {
+                    $data = $this->db->get($ckey);
+                }
+                else if ($this->type == 'apc') {
+                    $data = apc_fetch($ckey);
+                }
+
+                if ($this->debug) {
+                    $this->debug('get', $ckey, $data);
+                }
             }
 
             if ($data) {
@@ -284,13 +290,12 @@ class rcube_cache
         }
         else {
             $sql_result = $this->db->limitquery(
-                "SELECT data, cache_key".
-                " FROM " . $this->table.
-                " WHERE user_id = ?".
-                " AND cache_key = ?".
+                "SELECT `data`, `cache_key`".
+                " FROM {$this->table}".
+                " WHERE `user_id` = ? AND `cache_key` = ?".
                 // for better performance we allow more records for one key
                 // get the newer one
-                " ORDER BY created DESC",
+                " ORDER BY `created` DESC",
                 0, 1, $this->userid, $this->prefix.'.'.$key);
 
             if ($sql_arr = $this->db->fetch_assoc($sql_result)) {
@@ -315,7 +320,6 @@ class rcube_cache
         return $this->cache[$key];
     }
 
-
     /**
      * Writes single cache record into DB.
      *
@@ -337,7 +341,22 @@ class rcube_cache
         }
 
         if ($this->type == 'memcache' || $this->type == 'apc') {
-            return $this->add_record($this->ckey($key), $data);
+            $result = $this->add_record($this->ckey($key), $data);
+
+            // make sure index will be updated
+            if ($result) {
+                if (!array_key_exists($key, $this->cache_sums)) {
+                    $this->cache_sums[$key] = true;
+                }
+
+                $this->load_index();
+
+                if (!$this->index_changed && !in_array($key, $this->index)) {
+                    $this->index_changed = true;
+                }
+            }
+
+            return $result;
         }
 
         $key_exists = array_key_exists($key, $this->cache_sums);
@@ -346,9 +365,8 @@ class rcube_cache
         // Remove NULL rows (here we don't need to check if the record exist)
         if ($data == 'N;') {
             $this->db->query(
-                "DELETE FROM " . $this->table.
-                " WHERE user_id = ?".
-                " AND cache_key = ?",
+                "DELETE FROM {$this->table}".
+                " WHERE `user_id` = ? AND `cache_key` = ?",
                 $this->userid, $key);
 
             return true;
@@ -357,12 +375,12 @@ class rcube_cache
         // update existing cache record
         if ($key_exists) {
             $result = $this->db->query(
-                "UPDATE " . $this->table.
-                " SET created = " . $this->db->now().
-                    ", expires = " . ($this->ttl ? $this->db->now($this->ttl) : 'NULL').
-                    ", data = ?".
-                " WHERE user_id = ?".
-                " AND cache_key = ?",
+                "UPDATE {$this->table}".
+                " SET `created` = " . $this->db->now().
+                    ", `expires` = " . ($this->ttl ? $this->db->now($this->ttl) : 'NULL').
+                    ", `data` = ?".
+                " WHERE `user_id` = ?".
+                " AND `cache_key` = ?",
                 $data, $this->userid, $key);
         }
         // add new cache record
@@ -370,15 +388,14 @@ class rcube_cache
             // for better performance we allow more records for one key
             // so, no need to check if record exist (see rcube_cache::read_record())
             $result = $this->db->query(
-                "INSERT INTO " . $this->table.
-                " (created, expires, user_id, cache_key, data)".
+                "INSERT INTO {$this->table}".
+                " (`created`, `expires`, `user_id`, `cache_key`, `data`)".
                 " VALUES (" . $this->db->now() . ", " . ($this->ttl ? $this->db->now($this->ttl) : 'NULL') . ", ?, ?, ?)",
                 $this->userid, $key, $data);
         }
 
         return $this->db->affected_rows($result);
     }
-
 
     /**
      * Deletes the cache record(s).
@@ -399,114 +416,113 @@ class rcube_cache
             // Remove all keys
             if ($key === null) {
                 foreach ($this->index as $key) {
-                    $this->delete_record($key, false);
+                    $this->delete_record($this->ckey($key));
                 }
+
                 $this->index = array();
             }
             // Remove keys by name prefix
             else if ($prefix_mode) {
-                foreach ($this->index as $k) {
+                foreach ($this->index as $idx => $k) {
                     if (strpos($k, $key) === 0) {
-                        $this->delete_record($k);
+                        $this->delete_record($this->ckey($k));
+                        unset($this->index[$idx]);
                     }
                 }
             }
             // Remove one key by name
             else {
-                $this->delete_record($key);
+                $this->delete_record($this->ckey($key));
+                if (($idx = array_search($key, $this->index)) !== false) {
+                    unset($this->index[$idx]);
+                }
             }
+
+            $this->index_changed = true;
 
             return;
         }
 
         // Remove all keys (in specified cache)
         if ($key === null) {
-            $where = " AND cache_key LIKE " . $this->db->quote($this->prefix.'.%');
+            $where = " AND `cache_key` LIKE " . $this->db->quote($this->prefix.'.%');
         }
         // Remove keys by name prefix
         else if ($prefix_mode) {
-            $where = " AND cache_key LIKE " . $this->db->quote($this->prefix.'.'.$key.'%');
+            $where = " AND `cache_key` LIKE " . $this->db->quote($this->prefix.'.'.$key.'%');
         }
         // Remove one key by name
         else {
-            $where = " AND cache_key = " . $this->db->quote($this->prefix.'.'.$key);
+            $where = " AND `cache_key` = " . $this->db->quote($this->prefix.'.'.$key);
         }
 
         $this->db->query(
-            "DELETE FROM " . $this->table.
-            " WHERE user_id = ?" . $where,
+            "DELETE FROM {$this->table} WHERE `user_id` = ?" . $where,
             $this->userid);
     }
-
 
     /**
      * Adds entry into memcache/apc DB.
      *
-     * @param string  $key   Cache key name
-     * @param mxied   $data  Serialized cache data
-     * @param bollean $index Enables immediate index update
+     * @param string $key  Cache key name
+     * @param mixed  $data Serialized cache data
      *
      * @param boolean True on success, False on failure
      */
-    private function add_record($key, $data, $index=false)
+    private function add_record($key, $data)
     {
         if ($this->type == 'memcache') {
             $result = $this->db->replace($key, $data, MEMCACHE_COMPRESSED, $this->ttl);
-            if (!$result)
+
+            if (!$result) {
                 $result = $this->db->set($key, $data, MEMCACHE_COMPRESSED, $this->ttl);
+            }
         }
         else if ($this->type == 'apc') {
-            if (apc_exists($key))
+            if (apc_exists($key)) {
                 apc_delete($key);
+            }
+
             $result = apc_store($key, $data, $this->ttl);
         }
 
-        // Update index
-        if ($index && $result) {
-            $this->load_index();
-
-            if (array_search($key, $this->index) === false) {
-                $this->index[] = $key;
-                $data = serialize($this->index);
-                $this->add_record($this->ikey(), $data);
-            }
+        if ($this->debug) {
+            $this->debug('set', $key, $data, $result);
         }
 
         return $result;
     }
 
-
     /**
      * Deletes entry from memcache/apc DB.
+     *
+     * @param string $key Cache key name
+     *
+     * @param boolean True on success, False on failure
      */
-    private function delete_record($key, $index=true)
+    private function delete_record($key)
     {
         if ($this->type == 'memcache') {
             // #1488592: use 2nd argument
-            $this->db->delete($this->ckey($key), 0);
+            $result = $this->db->delete($key, 0);
         }
         else {
-            apc_delete($this->ckey($key));
+            $result = apc_delete($key);
         }
 
-        if ($index) {
-            if (($idx = array_search($key, $this->index)) !== false) {
-                unset($this->index[$idx]);
-            }
+        if ($this->debug) {
+            $this->debug('delete', $key, null, $result);
         }
+
+        return $result;
     }
-
 
     /**
      * Writes the index entry into memcache/apc DB.
      */
     private function write_index()
     {
-        if (!$this->db) {
-            return;
-        }
-
-        if ($this->type == 'db') {
+        if (!$this->db || $this->type == 'db') {
             return;
         }
 
@@ -514,10 +530,15 @@ class rcube_cache
 
         // Make sure index contains new keys
         foreach ($this->cache as $key => $value) {
-            if ($value !== null) {
-                if (array_search($key, $this->index) === false) {
-                    $this->index[] = $key;
-                }
+            if ($value !== null && !in_array($key, $this->index)) {
+                $this->index[] = $key;
+            }
+        }
+
+        // new keys added using self::write()
+        foreach ($this->cache_sums as $key => $value) {
+            if ($value === true && !in_array($key, $this->index)) {
+                $this->index[] = $key;
             }
         }
 
@@ -525,13 +546,12 @@ class rcube_cache
         $this->add_record($this->ikey(), $data);
     }
 
-
     /**
      * Gets the index entry from memcache/apc DB.
      */
     private function load_index()
     {
-        if (!$this->db) {
+        if (!$this->db || $this->type == 'db') {
             return;
         }
 
@@ -540,6 +560,7 @@ class rcube_cache
         }
 
         $index_key = $this->ikey();
+
         if ($this->type == 'memcache') {
             $data = $this->db->get($index_key);
         }
@@ -547,9 +568,12 @@ class rcube_cache
             $data = apc_fetch($index_key);
         }
 
+        if ($this->debug) {
+            $this->debug('get', $index_key, $data);
+        }
+
         $this->index = $data ? unserialize($data) : array();
     }
-
 
     /**
      * Creates per-user cache key name (for memcache and apc)
@@ -562,7 +586,6 @@ class rcube_cache
     {
         return sprintf('%d:%s:%s', $this->userid, $this->prefix, $key);
     }
-
 
     /**
      * Creates per-user index cache key name (for memcache and apc)
@@ -608,8 +631,10 @@ class rcube_cache
             $this->max_packet = 2097152; // default/max is 2 MB
 
             if ($this->type == 'db') {
-                $value = $this->db->get_variable('max_allowed_packet', 1048500);
-                $this->max_packet = min($value, $this->max_packet) - 2000;
+                if ($value = $this->db->get_variable('max_allowed_packet', $this->max_packet)) {
+                    $this->max_packet = $value;
+                }
+                $this->max_packet -= 2000;
             }
             else if ($this->type == 'memcache') {
                 $stats = $this->db->getStats();
@@ -623,5 +648,19 @@ class rcube_cache
         }
 
         return $this->max_packet;
+    }
+
+    /**
+     * Write memcache/apc debug info to the log
+     */
+    private function debug($type, $key, $data = null, $result = null)
+    {
+        $line = strtoupper($type) . ' ' . $key;
+
+        if ($data !== null) {
+            $line .= ' ' . ($this->packed ? $data : serialize($data));
+        }
+
+        rcube::debug($this->type, $line, $result);
     }
 }
